pyproject.toml
[project]
    authors = [{ name = "Nate Landau", email = "github@natelandau.com" }]
    dependencies = [
        "authlib>=1.6.9",
        "flask-caching>=2.3.1",
        "flask-debugtoolbar>=0.16.0",
        "flask-session>=0.8.0",
        "flask-talisman>=1.1.0",
        "flask-wtf>=1.2.2",
        "flask>=3.1.3",
        "gunicorn>=25.3.0",
        "humanize>=4.15.0",
        "jinja2-fragments>=1.12.0",
        "jinjax>=0.63",
        "loguru>=0.7.3",
        "markdown2>=2.5.5",
        "pydantic-settings>=2.13.1",
        "redis>=7.4.0",
        "requests>=2.33.1",
        "valentina-python-client[testing]>=1.20.1",
    ]
    description = "Web client for the Valentina Noir API."
    license = { file = "LICENSE" }
    name = "vweb"
    readme = "README.md"
    requires-python = ">=3.13,<3.14"
    version = "0.1.0"

    [project.scripts]
        vweb = "vweb:main"

[build-system]
    build-backend = "uv_build"
    requires      = ["uv_build>=0.10.4,<0.11.0"]

[dependency-groups]
    dev = [
        "commitizen>=4.13.9",
        "djlint>=1.36.4",
        "duty>=1.9.0",
        "prek>=0.3.8",
        "pytest-clarity>=1.0.1",
        "pytest-cov>=7.1.0",
        "pytest-devtools>=1.0.0",
        "pytest-mock>=3.15.1",
        "pytest-repeat>=0.9.4",
        "pytest-sugar>=1.1.1",
        "pytest-timeout>=2.4.0",
        "pytest-xdist>=3.8.0",
        "ruff>=0.15.9",
        "shellcheck-py>=0.11.0.1",
        "ty>=0.0.29",
        "typos>=1.45.0",
        "yamllint>=1.38.0",
    ]

[tool.commitizen]
    bump_message               = "bump(release): v$current_version → v$new_version"
    changelog_merge_prerelease = true
    tag_format                 = "v$version"
    update_changelog_on_bump   = true
    version_files              = ["src/vweb/__init__.py:__version__"]
    version_provider           = "uv"

[tool.coverage.report] # https://coverage.readthedocs.io/en/latest/config.html#report
    exclude_lines = [
        'def __repr__',
        'except [\w\s\._]+ as .*:',
        'if TYPE_CHECKING',
        'pragma: no cover',
        'raise Abort',
        'raise Exit',
    ]
    fail_under = 0
    precision = 1
    show_missing = true
    skip_covered = true
    skip_empty = true

[tool.coverage.run]
    branch       = true
    command_line = "--module pytest"
    data_file    = ".cache/coverage"
    omit         = ["tests/*"]
    source       = ["src"]

[tool.coverage.xml]
    output = ".cache/coverage.xml"

[tool.djlint]
    extend_exclude = ".vscode, .github, .git, .cache, __pycache__, .venv, tests, .claude, docs"
    ignore         = "H030,H031,H021,H006,H013"
    # ignore_case    = true
    blank_line_after_tag  = "load,extends,include,set"
    blank_line_before_tag = "load,extends,include,set"
    # max_attribute_length  = 10
    # format_attribute_template_tags = false
    indent          = 2
    max_blank_lines = 1
    max_line_length = 72
    profile         = "jinja"
    use_gitignore   = true

[tool.pytest.ini_options]

    addopts = "--color=yes --doctest-modules --exitfirst --failed-first --strict-config --strict-markers"
    cache_dir = ".cache/pytest"
    filterwarnings = [
        'error',
        'ignore:.*Pydantic.*:UserWarning',
        'ignore::DeprecationWarning:',
        'ignore::pytest.PytestUnraisableExceptionWarning',
        'ignore:unclosed.*:ResourceWarning',
    ]
    markers = [
        "drop_db: drops the database before the test creating fresh data",
        "no_db: marks tests as not needing a database",
        "serial",
    ]
    testpaths = ["tests"]
    timeout = 10
    xfail_strict = true

[tool.ruff] # https://github.com/charliermarsh/ruff

    exclude        = [".cache", ".git", ".venv", "_build", "build", "dist"]
    fix            = true
    line-length    = 100
    output-format  = "grouped"
    src            = ["src", "tests"]
    target-version = "py313"
    [tool.ruff.lint]
        ignore = [
            "ANN002",  # Missing type annotation for `*args`
            "ANN003",  # Missing type annotation for `**kwargs`
            "ANN204",  # missing return type annotation for special method `__init__`
            "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed,
            "B006",    # mutable-argument-default
            "B008",    # function-call-in-default-argument
            "COM812",  # Trailing comma missing"
            "CPY001",  # Missing copyright notice at top of file
            "D107",    # undocumented-public-init
            "E501",    # line-too-long
            "PLC0415", # Imports should be at top of file
            # "FBT001", # Boolean-typed positional argument in function definition
            # "FBT002", # Boolean-typed positional argument in function definition
            "FIX002", # Line contains TODO, consider resolving the issue
            "S311",   # suspicious-non-cryptographic-random-usage
            "TD001",  # invalid-todo-tag
            "TD002",  # Missing author in TODO
            "TD003",  # Missing issue link on the line following this TODO
        ]
        per-file-ignores = { "tests/**/*.py" = [
            "A002",
            "A003",
            "ANN001",  # Missing type annotation for function argument `cls`
            "ANN002",  # Missing type annotation for `*args`
            "ANN003",  # Missing type annotation for `**kwargs`
            "ANN201",  # Missing return type annotation
            "ARG001",  # Unused function argument
            "ARG002",  # Unused method argument
            "ARG005",  # Unused lambda argument
            "D102",
            "ERA001",  # Commented out code
            "F403",
            "F405",    # May be undefined from type imports
            "PGH003",  # Use specific rule codes when ignoring type issues
            "PLC0415", # Imports should be at top of file
            "PLR0913",
            "PLR2004",
            "S101",
            "SLF001",  # Calling private method
            "W292",    # No blank line at end of file - included b/c cursor struggles to add a trailing newline and burns through requests trying to fix this linting error.
        ] }
        select = ["ALL"]
        unfixable = [
            "ERA001", # Commented out code
            "F401",   # unused-import
            "F841",   # unused-variable
        ]

        [tool.ruff.lint.mccabe]
            # Unlike Flake8, default to a complexity level of 10.
            max-complexity = 10

        [tool.ruff.lint.pydocstyle]
            convention = "google"

        [tool.ruff.lint.pylint]
            max-args = 6

    [tool.ruff.format]
        indent-style              = "space"
        line-ending               = "auto"
        quote-style               = "double"
        skip-magic-trailing-comma = false
